remove cyclic dep between MongoDb and managefile
